+++ src/lib/utils/project.ts
@@ -31,7 +31,7 @@ export async function commitVersion(version: string, projectRoot?: string): Prom
 export async function release(version: string, locations: ProjectOptions): Promise<void> {
   await Promise.all([
     assertUnusedTag(getVersionTag(version)),
-    git.assertCleanMaster()
+    git.assertCleanBranch(["master", getVersionTag(version)])
   ]);
   await setPackageVersion(version, locations);
   await commitVersion(version, locations.root);
@@ -63,8 +63,10 @@ export async function setPackageVersion(version: string, locations: ProjectOptio
   return writePackage(packageData, locations);
 }
 
-export async function getNextVersion(bumpKind: "major" | "minor" | "patch",
-                                     locations: ProjectOptions): Promise<string> {
+export async function getNextVersion(
+  bumpKind: "major" | "minor" | "patch",
+  locations: ProjectOptions
+): Promise<string> {
   const packageData: PackageJson = await readPackage(locations);
   return semver.inc(packageData.version, bumpKind);
 }