+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Review the organization of the build targets
 - Use `displayName` on task functions, instead of separate `generateTaskName` functions.
 
-## Next
+## 0.14.0 (2017-05-10)
 
 - **[Breaking]** Enforce trailing commas by default for multiline objects
 - **[Feature]** Allow bump from either `master` or a branch with the same name as the tag (exampel: `v1.2.3`)